chore: reorder and clean all files
