+++ src/expand_utils.c
@@ -1,16 +1,63 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   clean_utils.c                                      :+:      :+:    :+:   */
+/*   expand_utils.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/08/15 22:56:08 by pmarkaid          #+#    #+#             */
-/*   Updated: 2024/08/18 15:55:51 by pmarkaid         ###   ########.fr       */
+/*   Created: 2024/08/22 20:53:07 by pmarkaid          #+#    #+#             */
+/*   Updated: 2024/08/22 20:55:07 by pmarkaid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+
+bool	is_inside_double_quotes(char *str, int index)
+{
+	bool	inside_double_quotes;
+	int		i;
+
+	inside_double_quotes = false;
+	i = 0;
+	while (i <= index && str[i] != '\0')
+	{
+		if (str[i] == '\"')
+			inside_double_quotes = !inside_double_quotes;
+		i++;
+	}
+	return (inside_double_quotes);
+}
+
+bool	envir_must_be_expanded(char *str, int index)
+{
+	if (!is_in_quote(str, index))
+		return (true);
+	if (is_inside_double_quotes(str, index))
+		return (true);
+	return (false);
+}
+
+bool	is_in_quote(char *str, int index)
+{
+	bool	inside_single_quotes;
+	bool	inside_double_quotes;
+	int		i;
+
+	inside_single_quotes = false;
+	inside_double_quotes = false;
+	i = 0;
+	while (i <= index && str[i] != '\0')
+	{
+		if (str[i] == '\'')
+			inside_single_quotes = !inside_single_quotes;
+		if (str[i] == '\"')
+			inside_double_quotes = !inside_double_quotes;
+		i++;
+	}
+	if (inside_single_quotes || inside_double_quotes)
+		return (true);
+	return (false);
+}
 
 char	*get_envir_name(char *str)
 {
@@ -39,51 +86,4 @@ char	*get_envir_value(char *str, t_macro *macro)
 	if (!envir_value)
 		return (NULL);
 	return (envir_value);
-}
-
-bool	is_inside_double_quotes(char *str, int index)
-{
-	bool	inside_double_quotes;
-	int		i;
-
-	inside_double_quotes = false;
-	i = 0;
-	while (i <= index && str[i] != '\0')
-	{
-		if (str[i] == '\"')
-			inside_double_quotes = !inside_double_quotes;
-		i++;
-	}
-	return (inside_double_quotes);
-}
-
-bool	is_in_quote(char *str, int index)
-{
-	bool	inside_single_quotes;
-	bool	inside_double_quotes;
-	int		i;
-
-	inside_single_quotes = false;
-	inside_double_quotes = false;
-	i = 0;
-	while (i <= index && str[i] != '\0')
-	{
-		if (str[i] == '\'')
-			inside_single_quotes = !inside_single_quotes;
-		if (str[i] == '\"')
-			inside_double_quotes = !inside_double_quotes;
-		i++;
-	}
-	if (inside_single_quotes || inside_double_quotes)
-		return (true);
-	return (false);
-}
-
-bool	envir_must_be_expanded(char *str, int index)
-{
-	if (!is_in_quote(str, index))
-		return (true);
-	if (is_inside_double_quotes(str, index))
-		return (true);
-	return (false);
 }
